+++ app/shared/availability-view/sidebar/group-info/resource-info/__tests__/ResourceInfoContainer.spec.js
@@ -3,7 +3,7 @@ import { Link } from 'react-router-dom';
 
 import { shallowWithIntl } from '../../../../../../utils/testUtils';
 import { UnconnectedResourceInfo as ResourceInfo, selector } from '../ResourceInfoContainer';
-import UnpublishedLabel from '../../../../../label/un-published/UnpublishedLabel';
+import UnpublishedLabel from '../../../../../label/unpublished/UnpublishedLabel';
 
 function getState() {
   return {
